requirements.txt
streamlit
langchain
langchain-groq
langchain-mcp-adapters
langchain-classic
python-dotenv
nest_asyncio